+++ required_providers.tf
@@ -3,17 +3,17 @@
 #
 
 # -
-# - Required Terraform providers & versions
+# - Required Terraform providers + versions
 # -
 terraform {
   required_providers {
     random = {
       source  = "hashicorp/random"
-      version = ">=3.6"
+      version = "~> 3.6"
     }
     time = {
       source  = "hashicorp/time"
-      version = ">=0.12"
+      version = "~> 0.12"
     }
   }
 }